Various styling enhancements
